build: Rework editables for full PEP-660 support
The previously used mechanism of synthezising a package using hatchling sort of works, and was enough to prove the concept of editables with Nix.

This changes `mkDerivationEditable` to use PEP-660 and should support pretty much everything.

The big complication in this development was the need for source level patching since editables are a pretty loose concept and are implemented differently by different build systems.
Source patching uses https://github.com/Instagram/LibCST/.
I'd like to rewrite this bit purely in Rust later.

So far I've successfully used this support with a few build systems, most notably meson.

For meson this requires manually calling `patch_editable.py` to create the initial files.
After that point dynamic recompilation "just works".
This will be exposed as a separate package somehow, just not yet.