Allow Cargo.lock hash to be provided when using importCargoLock
Add optional cargoLockHash parameter that makes Cargo.lock extraction a fixed-output derivation.

Use pkgsBuildHost rather than pkgs for build -- as a FOD, this doesn't change the build hash.

This allows platforms that are doing cross-evalution but not cross-compilation to generate identical derivation hashes between cross-evaluation and local build, so long as they (1) provide a hash for Cargo.lock files, and (2) ensure that `pkgsBuildHost` is set to reflect the system doing the evaluation.